requirements.txt
pandas
numpy
scikit-learn
xgboost
tqdm
joblib
flask
pymongo
python-whois
requests
reportlab
matplotlib
seaborn
dnspython
